requirements.txt
openai
pandas
python-dotenv
yagmail
apscheduler
GitPython
flask
resend
